Revert "Entire refactoring + stream implementation"
This reverts commit 1ecf024ee88725af624bdff78134fc89fb2dd2f0.

Explanation here: https://github.com/axel-op/feature_discovery/pull/17#issuecomment-528336106
