+++ example/lib/main.dart
@@ -100,12 +100,8 @@ class _MyHomePageState extends State<MyHomePage> {
                         .textTheme
                         .button
                         .copyWith(color: Colors.white)),
-                onPressed: () => FeatureDiscovery.completeCurrentStep(context),
-              ),
-              FlatButton(
-                padding: const EdgeInsets.all(0),
-                child: Text('Cancel'),
-                onPressed: () => FeatureDiscovery.dismissCurrentStep(context),
+                onPressed: () =>
+                    FeatureDiscovery.markStepComplete(context, feature1),
               )
             ],
           ),
