+++ lib/feature_discovery.dart
@@ -2,6 +2,4 @@ library feature_discovery;
 
 //  export 'src/layout.dart';
 export 'src/feature_discovery.dart';
-export 'src/described_feature_overlay.dart';
 export 'src/ensure_visible.dart';
-export 'src/enums.dart';
