--- lib/src/enums.dart
@@ -1,18 +1,0 @@
-enum DescribedFeatureContentOrientation {
-  above,
-  below,
-}
-
-enum OverlayState {
-  closed,
-  opening,
-  pulsing,
-  activating,
-  dismissing,
-}
-
-enum ContentOrientation {
-  above,
-  below,
-  trivial,
-}
